desktop: Separate menu bar into a module
